project/requirements.txt
numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
sympy>=1.13
PyYAML>=6.0
tqdm>=4.66
openai>=1.40
sentence-transformers>=3.0
joblib>=1.4
